🚚 Refactor(characters): Rename files
